Tests/LibWeb/Text/expected/Fetch/fetch-later-api-shape.txt

typeof fetchLater: function
result instanceof FetchLaterResult: true
typeof result.activated: boolean
result.activated is readonly: true
PASS
